Add VScript support for TF objects, allow one->many teleporters
A number of VScript functions were added to the various building classes. SetBuilder was modified to make sure m_aObjects in CTFPlayer and CTFTeam would be updated.

Teleporters got one->many support for the fun of it. Only one teleporter can be active at a time, but the matching teleporter can be changed and both will link.